PANEL_for_chat/requirements.txt
gradio==4.19.2
requests==2.31.0
python-docx==1.0.1
python-pptx==0.6.23
pdfplumber==0.10.3
Pillow==10.2.0
openai==1.12.0
python-dotenv==1.0.1
markdown==3.5.2 